+++ react/stisla-page/dashboard/general/readme.md
@@ -1,0 +1,12 @@
+# Use Statistic Card 4
+
+```html
+  <div className="row">
+    <StatisticCard4 
+      header="Total Admin"
+      text="10"
+      icon="far fa-user"
+      iconBg="bg-primary"/>
+  </div>
+```
+1. Bungkus StatisticCard dengan class Row
